+++ scripts/examples/simple.py
@@ -1,7 +1,7 @@
 from private.private import start, task, arg, group
 
 @task
-@arg('arg1', type=int, desc= 'Some value.')
+@arg('arg1', type=int, desc= 'Value for arg1')
 @arg('arg2', type=int)
 def task1(arg1, arg2=1):
   print arg1, arg2